[DRAFT] Implement db_datareader/db_datawriter

+++ contrib/babelfishpg_tsql/src/dbcmds.c
@@ -51,12 +51,16 @@ static List *gen_createdb_subcmds(const char *schema,
 								  const char *dbo,
 								  const char *db_owner,
 								  const char *guest,
-								  const char *guest_schema);
+								  const char *guest_schema,
+								  const char *db_datareader,
+								  const char *db_datawriter);
 static List *gen_dropdb_subcmds(const char *schema,
 								const char *db_owner,
 								const char *dbo,
 								List *db_users,
-								const char *guest_schema);
+								const char *guest_schema,
+								const char *db_datareader,
+								const char *db_datawriter);
 static Oid	do_create_bbf_db(ParseState *pstate, const char *dbname, List *options, const char *owner);
 static void create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, const char *owner, int16 dbid);
 static void drop_related_bbf_namespace_entries(int16 dbid);
@@ -80,7 +84,8 @@ get_sys_babelfish_db_seq_oid()
  * Generate subcmds for CREATE DATABASE. Note 'guest' can be NULL.
  */
 static List *
-gen_createdb_subcmds(const char *schema, const char *dbo, const char *db_owner, const char *guest, const char *guest_schema)
+gen_createdb_subcmds(const char *schema, const char *dbo, const char *db_owner, const char *guest, const char *guest_schema,
+					const char *db_datareader, const char *db_datawriter)
 {
 	StringInfoData query;
 	List	   *res;
@@ -95,6 +100,8 @@ gen_createdb_subcmds(const char *schema, const char *dbo, const char *db_owner, 
 	 */
 	initStringInfo(&query);
 
+	appendStringInfo(&query, "CREATE ROLE dummy INHERIT; ");
+	appendStringInfo(&query, "CREATE ROLE dummy INHERIT; ");
 	appendStringInfo(&query, "CREATE ROLE dummy CREATEROLE INHERIT; ");
 	appendStringInfo(&query, "CREATE ROLE dummy INHERIT CREATEROLE ROLE sysadmin IN ROLE dummy; ");
 	appendStringInfo(&query, "GRANT CREATE, CONNECT, TEMPORARY ON DATABASE dummy TO dummy; ");
@@ -118,9 +125,9 @@ gen_createdb_subcmds(const char *schema, const char *dbo, const char *db_owner, 
 	res = raw_parser(query.data, RAW_PARSE_DEFAULT);
 
 	if (guest)
-		expected_stmt_num = list_length(logins) > 0 ? 9 : 8;
+		expected_stmt_num = list_length(logins) > 0 ? 11 : 10;
 	else
-		expected_stmt_num = 6;
+		expected_stmt_num = 8;
 
 	if (list_length(res) != expected_stmt_num)
 		ereport(ERROR,
@@ -129,6 +136,12 @@ gen_createdb_subcmds(const char *schema, const char *dbo, const char *db_owner, 
 						expected_stmt_num, list_length(res))));
 
 	/* Replace dummy elements in parsetree with real values */
+	stmt = parsetree_nth_stmt(res, i++);
+	update_CreateRoleStmt(stmt, db_datareader, NULL, NULL);
+
+	stmt = parsetree_nth_stmt(res, i++);
+	update_CreateRoleStmt(stmt, db_datawriter, NULL, NULL);
+
 	stmt = parsetree_nth_stmt(res, i++);
 	update_CreateRoleStmt(stmt, db_owner, NULL, NULL);
 
@@ -181,13 +194,15 @@ gen_dropdb_subcmds(const char *schema,
 				   const char *db_owner,
 				   const char *dbo,
 				   List *db_users,
-				   const char *guest_schema)
+				   const char *guest_schema,
+				   const char *db_datareader,
+				   const char *db_datawriter)
 {
 	StringInfoData query;
 	List	   *stmt_list;
 	ListCell   *elem;
 	Node	   *stmt;
-	int			expected_stmts = 6;
+	int			expected_stmts = 8;
 	int			i = 0;
 
 	initStringInfo(&query);
@@ -198,7 +213,7 @@ gen_dropdb_subcmds(const char *schema,
 	{
 		char	   *user_name = (char *) lfirst(elem);
 
-		if (strcmp(user_name, db_owner) != 0 && strcmp(user_name, dbo) != 0)
+		if (strcmp(user_name, db_owner) != 0 && strcmp(user_name, dbo) != 0 && strcmp(user_name, db_datareader) != 0 && strcmp(user_name, db_datawriter) != 0)
 		{
 			appendStringInfo(&query, "DROP OWNED BY dummy CASCADE; ");
 			appendStringInfo(&query, "DROP ROLE dummy; ");
@@ -208,6 +223,8 @@ gen_dropdb_subcmds(const char *schema,
 	/* Then drop db_owner and dbo in that order */
 	appendStringInfo(&query, "DROP OWNED BY dummy, dummy CASCADE; ");
 	appendStringInfo(&query, "REVOKE CREATE, CONNECT, TEMPORARY ON DATABASE dummy FROM dummy; ");
+	appendStringInfo(&query, "DROP ROLE dummy; ");
+	appendStringInfo(&query, "DROP ROLE dummy; ");
 	appendStringInfo(&query, "DROP ROLE dummy; ");
 	appendStringInfo(&query, "DROP ROLE dummy; ");
 
@@ -229,7 +246,7 @@ gen_dropdb_subcmds(const char *schema,
 	{
 		char	   *user_name = (char *) lfirst(elem);
 
-		if (strcmp(user_name, db_owner) != 0 && strcmp(user_name, dbo) != 0)
+		if (strcmp(user_name, db_owner) != 0 && strcmp(user_name, dbo) != 0 && strcmp(user_name, db_datareader) != 0 && strcmp(user_name, db_datawriter) != 0)
 		{
 			stmt = parsetree_nth_stmt(stmt_list, i++);
 			update_DropOwnedStmt(stmt, list_make1(user_name));
@@ -249,6 +266,11 @@ gen_dropdb_subcmds(const char *schema,
 	update_DropRoleStmt(stmt, db_owner);
 	stmt = parsetree_nth_stmt(stmt_list, i++);
 	update_DropRoleStmt(stmt, dbo);
+
+	stmt = parsetree_nth_stmt(stmt_list, i++);
+	update_DropRoleStmt(stmt, db_datareader);
+	stmt = parsetree_nth_stmt(stmt_list, i++);
+	update_DropRoleStmt(stmt, db_datawriter);
 
 	return stmt_list;
 }
@@ -406,6 +428,8 @@ create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, co
 	const char *dbo_scm;
 	const char *dbo_role;
 	const char *db_owner_role;
+	const char *db_datareader;
+	const char *db_datawriter;
 	const char *guest_scm;
 	NameData	default_collation;
 	NameData	owner_namedata;
@@ -483,6 +507,8 @@ create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, co
 	dbo_scm = get_dbo_schema_name(dbname);
 	dbo_role = get_dbo_role_name(dbname);
 	db_owner_role = get_db_owner_name(dbname);
+	db_datareader = get_db_datareader_name(dbname);
+	db_datawriter = get_db_datawriter_name(dbname);
 	guest = get_guest_role_name(dbname);
 	guest_scm = get_guest_schema_name(dbname);
 
@@ -539,7 +565,7 @@ create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, co
 	/* Advance cmd counter to make the database visible */
 	CommandCounterIncrement();
 
-	parsetree_list = gen_createdb_subcmds(dbo_scm, dbo_role, db_owner_role, guest, guest_scm);
+	parsetree_list = gen_createdb_subcmds(dbo_scm, dbo_role, db_owner_role, guest, guest_scm, db_datareader, db_datawriter);
 
 	GetUserIdAndSecContext(&save_userid, &save_sec_context);
 	old_createrole_self_grant = pstrdup(GetConfigOption("createrole_self_grant", false, true));
@@ -580,9 +606,9 @@ create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, co
 			wrapper->stmt_location = 0;
 			stmt_number++;
 			if (guest && list_length(parsetree_list) == stmt_number)
-				wrapper->stmt_len = 19;
+				wrapper->stmt_len = 21;
 			else
-				wrapper->stmt_len = 18;
+				wrapper->stmt_len = 20;
 
 			/* do this step */
 			ProcessUtility(wrapper,
@@ -604,6 +630,10 @@ create_bbf_db_internal(ParseState *pstate, const char *dbname, List *options, co
 			add_to_bbf_authid_user_ext(dbo_role, "dbo", dbname, "dbo", NULL, false, true, false);
 		if (db_owner_role)
 			add_to_bbf_authid_user_ext(db_owner_role, "db_owner", dbname, NULL, NULL, true, true, false);
+		if (db_datareader)
+			add_to_bbf_authid_user_ext(db_datareader, "db_datareader", dbname, NULL, NULL, true, true, false);
+		if (db_datawriter)
+			add_to_bbf_authid_user_ext(db_datawriter, "db_datawriter", dbname, NULL, NULL, true, true, false);
 		if (guest)
 		{
 			/*
@@ -635,6 +665,8 @@ drop_bbf_db(const char *dbname, bool missing_ok, bool force_drop)
 	int16		dbid;
 	const char *schema_name;
 	const char *db_owner_role;
+	const char *db_datareader;
+	const char *db_datawriter;
 	const char *dbo_role;
 	const char *guest_schema_name;
 	List	   *db_users_list;
@@ -724,6 +756,8 @@ drop_bbf_db(const char *dbname, bool missing_ok, bool force_drop)
 		schema_name = get_dbo_schema_name(dbname);
 		dbo_role = get_dbo_role_name(dbname);
 		db_owner_role = get_db_owner_name(dbname);
+		db_datareader = get_db_datareader_name(dbname);
+		db_datawriter = get_db_datawriter_name(dbname);
 		/* Get a list of all the database's users */
 		db_users_list = get_authid_user_ext_db_users(dbname);
 		guest_schema_name = get_guest_schema_name(dbname);
@@ -732,7 +766,9 @@ drop_bbf_db(const char *dbname, bool missing_ok, bool force_drop)
 											db_owner_role,
 											dbo_role,
 											db_users_list,
-											guest_schema_name);
+											guest_schema_name,
+											db_datareader,
+											db_datawriter);
 
 		/* Run all subcommands */
 		foreach(parsetree_item, parsetree_list)
